requirements.txt
Flask==3.1.0
nltk==3.9.1
numpy==2.2.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.1.0
pyttsx3==2.98
pywin32==308
tensorflow==2.18.0
tensorflow_intel==2.18.0
